Add files via upload
Jupyter Notebook code for calculating defect sizes from a (CG)MD trajectory. Output is a text file with the size of each defect. Sorting and binning is required for generating a defect size distribution. Default values for the bead size should be set before running the code, as they depend on the force field used.